requirements.txt
flask
markdown
python-dotenv
python-frontmatter
colorama
gunicorn